Adding locks
There are two locks, because there are two different kind of mutual exclusivity.

First deals with sessions, since we don't want to read/write sessions from multiple places, and we want acquire/release to be atomical.

Second deals with enumerating and call - because of some underlying issues, we cannot make calls and enumeration at the same time.

*However*, while doing call, we can reasonably assume that the enumeration returns similar values to previous enumeration. So, instead of blocking enumerate on call, we just set that call is hapenning and we use the previous USB info.