Try opening 3 times with delay
Because of how webUSB works, Google Chrome tries to read from the USB after connecting Trezor, to read metadata.

When the timing is wrong, it can happen that bridge is trying to open the device at the same time as Chrome. It's hard to replicate, but it happens once in a while.

To prevent throwing errors, it's better to try a few times before throwing out errors.