windows: Add runtime check for admin
`RequestExecutionLevel admin` actually does not work if the Windows machine has UAC turned off, and is only a "hint" for Windows to run the installer in admin mode.

This adds an additional check on runtime.

Copied from here, tested on Windows 7 (I cannot turn UAC off on Win10, maybe it is not possible).,

https://stackoverflow.com/questions/8821476/nsis-installing-an-application-to-always-run-as-administrator/8825202