usb: Do not finish read queue on disconnected
"Finishing reading queue" is necessary only because sometimes the handle is closed before the device sends all the USB packets, which happens only when one app "steals" device from another.

It is never necessary on disconnected devices; on the contrary, it sometimes causes issues.

This requires a slight change in the device interface.