libusb: Add cancel to sync.c, don't use timeouts
The main reason why we used interrupt with timeouts is that we cannot
cancel transfers, using the libusb synchronnous API.

However, the repeated timeouts cause issues with the bootloader for unknown reasons.

Instead of using timeouts, I added a function to cancel transfers to synch libusb API.
Then we can safely remove the timeouts from the code, since we cancel the calls manually.
(We cannot remove them completely - we still use them for finishing the read queue.)

What I did not know before is that using libusb close on a device with pending synch transfers
**does not cancel the pending transfers** - and the transfers then hang forever!
Also, libusb device close is a non-blocking function.

---

This is a little complex, but the other option would be to basically rewrite sync.c
into go.

Unfortunately this probably breaks FreeBSD again, since we add API to libusb.