libusb: Use device IDs as unique paths
Significant reworking of libusb enumeration and paths

So far, "path" was actually a USB port, but that could cause some issues when the device was disconnected and then reconnected again, when the service was not constantly watching uploads via listen. That was unfixable.

Instead of that, I have rewritten the libusb to use "paths" that are unique for each device and increase every device connection/disconnection. I compare devices by just comparing pointers of libusb lowlevel objects.

However there are 2 issues on windows:

1) we still have the other bug with multiple devices on windows, so we have to work around that

2) on windows, if the device is disconnected and connected again, or if it's two bootloader devices, the path does not increase.

First is solved, second is not that big of an issue I guess