api: Use unique device IDs instead of USB paths
Instead of exporting USB paths, export unique paths for each device.

This is being done by doing "internal enumerate" if at least 1 device is connected;
as long as it is disconnected, the next device on the same USB path will have a different ID

This prevents accidental call of acquire/ on reconnected devices on the same port.

This is a cleaner variant of https://github.com/trezor/trezord-go/pull/122 ; unlike that, it actually works with HID and does not rely on wonky pointer comparing. (This compares the USB paths, which should be more reliable.)

As a side-effect, it removes the prefixes from the paths, I hope that nothing from the ecosystem relies on that.

Fixes https://github.com/trezor/trezord-go/issues/121