api: create api/, use internally in server/
To help other devs to use trezord-go as a library, I have added an api/ folder, which can be now used as a native golang API/library to contact Trezor devices.

To "test" the API properly, I have also modified the server/ to use directly the API, and not the core/ calls.

api.New(..) now also initializes all the USB libraries, to make the use easier; that was originally in main.go

I have also cleaned up the main.go a bit by breaking it into more files, when I was already rewriting it significantly

There are some caveats.

* I have intentionally made the API by copying the current HTTP calls and not making it more go-like; for example, Listen could be putting channel of devices instead of requiring previous, etc; basically because I want to "test" the new API calls by using it in the HTTP server itself.
* There is still no documentation :)
* There is no protobuf encoding/decoding, the API requires the user to do it themselves
* There might be a problem with the USB initialization in api.New, since the user of the API might already have LibUSB initialized and libusb requires to be initialized just once.
* There is no check if trezord-go is already running as a server. We should detect that and call that through HTTP if it already runs.
* I have NOT tested it with other software that already used trezord-go in a weird way as a library. The only one that I know of is rfjakob/gocryptfs which used it via xaionaro-go/cryptoWallet. Someone should look at and rewrite xaionaro-go/cryptoWallet and rfjakob/gocryptfs and send them PRs before merging this.