usb: Cleaning up read queue when canceled
Fixes https://github.com/trezor/trezord-go/issues/154 , but the
resulting code is so complex that I am not sure it's wort this.

(Since it will probably have bugs because of the complexity.)

Tested on UDP and Trezor T2.

(Note that UDP emulator will hang, but that's not the issue of this fix.
If we want to fix the hang/the black screen, we would also need to send empty Initialize
after the read here.)