fix: close libusb config description memory leak
The config descriptors from the libusb c-library are not being properly
closed by the golang code.

On my PC, when testing with MetaMask, this causes an approximately
10 MB per hour memory leak
This memory leak is not detected via golang's  pprof, as it's in the
c-library.

Closing these config descriptors using `defer` seems to fix this memory
leak in my testing.