Simplify signing, enable notarization
I have simplified signing and enabled notarization in docker
with rcodesign (a Rust-based tool, fully open-source; but it does connect to Apple servers for notarization).

This tool enables notarization and signing fully in Docker on any OS,
so it should work in CLI too.

There is some prep-work that needs to be done on Mac OS side and on Apple website;
but after that, the signing and notarization works automatically.

It's not directly relevant to this commit, but I think you can also use it on Trezor Suite;
a cursory glance there tells me that Trezor Suit is using electron-notarize, which needs to be run on Mac OS;
rcodesign does not need that.

Note that we need to use ldflags -s because of this issue of rcodesign
https://github.com/indygreg/PyOxidizer/issues/636