Remove githash from Makefiles, move to go code
In go 1.18, go is "magically" putting git hash to binary itself, which can be read
in the code itself. So we don't need the `git parse` in Makefiles.

(I actually forgot to put the GITHASH magic to the osx build, but better to fix it properly.)

The module version is _not_ put there, as it's hard to tell from just git. So we still need
that there.