Build Linux binaries with musl
When I tried to build with "normal" glibc, I got weird errors on old/new glibc.

Instead of building with glibc, which limits us with compatibility and weird issues, let's build with
musl, which will always produce 100% static library.

(Musl is a libc alternative. The good thing is musl is statically linked
and can cross-compile relatively easily.)

The actual hard work is building all these musl cross-compilers;
luckily, Rust people already did that for Rust :) we can steal that and reuse.

Note that this all works on both AMD and ARM hosts, as both the musl cross-compilers images
and the go images are published for both.