Fix linux build for linux hosts
The issue is with new git, and Linux Docker hosts.

It's a mitigation against CVE-2022-24765, as the owner of the outside repo needs to be the same as the current user; and the user inside the docker container is root. (Go is running git to do the version stamping.)

Simple fix is to add a safe directory, as the CVE does not apply to us.