Clean various warnings in low-level libraries.
We are using old versions of libusb and hidapi that result in some warnings.

We could fix that by updating hidapi+libusb, but I am actually deadly afraid to do that (don't fix what works); I know what pain was to ensure that the USB layer really works.

So I just went through all the warnings, saw how the other libraries fixed them, and did the same.

The changes:

* on mac, kIOMasterPortDefault is deprecated; we can just use 0 (or, as libusb does, add new constant and use that), in both libusb and hidapi
* on windows, just use memcpy instead of useless strncpy in hidapi
* on linux, use POSIX header instead of "incorrect" header that is redirected
* and also remove unused libusb set_debug from go code

Note, this would all be solved if we used upstream versions of all the libraries, but again, I am super afraid to do that.