Clean makefiles, move GPG signing to release.sh
I have cleaned up the makefiles. I have also moved the GPG signing to
the individual releases and out of existing Makefile.

Also I read the GPG key out of GPG instead of hard-coding it.

The negative is that now we require quite a lot of keys to set up...
especially the macos now needs tons of keys.

I will try to document them all somewhere.