Remove devcon+libwdi for removing/installing drivers
We have included two third party tools - libwdi and devcon - just in order to uninstall
old drivers and re-installing them again, on installing Trezor Bridge.

This was ONLY needed for Windows 7. As their WinUSB support was a bit broken.

As we decided to drop Windows 7, we can delete devcon and libwdi, and remove
the voodoo from NSIS installer. And also remove the dockerfiles, building devcon+libwdi.

We can also remove the crazy NSIS function for doing a substring match.

If we decide to return Win7 support, we can always revert this commit :)